+++ src/arm64/intrinsics.h
@@ -1,0 +1,8 @@
+#ifndef SIMDJSON_ARM64_INTRINSICS_H
+#define SIMDJSON_ARM64_INTRINSICS_H
+
+// This should be the correct header whether
+// you use visual studio or other compilers.
+#include <arm_neon.h>
+
+#endif //  SIMDJSON_ARM64_INTRINSICS_H
