+++ src/haswell/stage1_find_marks.h
@@ -5,6 +5,8 @@
 
 #ifdef IS_X86_64
 
+#include "haswell/intrinsics.h"
+#include "haswell/bitmanipulation.h"
 #include "haswell/simd_input.h"
 #include "haswell/simdutf8check.h"
 #include "simdjson/stage1_find_marks.h"
