+++ src/westmere/intrinsics.h
@@ -1,0 +1,11 @@
+#ifndef SIMDJSON_WESTMERE_INTRINSICS_H
+#define SIMDJSON_WESTMERE_INTRINSICS_H
+
+
+#ifdef _MSC_VER
+#include <intrin.h> // visual studio
+#else
+#include <x86intrin.h> // elsewhere
+#endif
+
+#endif //  SIMDJSON_WESTMERE_INTRINSICS_H
